refactor(db): collapse remote module; move integration tests to tests/
Flatten crates/db/lib/remote/{mod.rs,integration.rs} into a single
remote.rs adapter file, and promote the live-sqld integration tests into
a proper tests/remote_libsql.rs integration target. Imports updated from
crate:: to microsandbox_db:: public paths.
